\name{dimsep}
\alias{dimsep}
\title{Seperate the Extracted Values}
\usage{
  dimsep(eres.values)
}
\arguments{
  \item{eres.values}{values of Lingo results from
  \code{\link{extractLingo}}.}
}
\value{
  A list which includes "dim1", "dim2", ...
}
\description{
  Seperate the extracted values of Lingo results by each
  dimmention.
}
\details{
  This function is used to seperate values of Lingo results
  into a list which can be easily used by R.
}

