\name{writeLingo}
\alias{writeLingo}
\title{Write Lingo Data}
\usage{
  writeLingo(datalist, ldt = "lingo.txt")
}
\arguments{
  \item{datalist}{the data list for the Lingo data file.}

  \item{ldt}{the Lingo data file.}
}
\value{
  A Lingo data file contains all data which can be imported
  by Lingo code.
}
\description{
  Write Lingo data into a text file.
}
\details{
  This function is used to convert the data list in R into
  a data file which can be imported by Lingo.
}

